[monitoring] Move check into monitoring #13
Closes #13
