--- openwisp_monitoring/check/models.py
@@ -1,9 +1,0 @@
-from swapper import swappable_setting
-
-from .base.models import AbstractCheck
-
-
-class Check(AbstractCheck):
-    class Meta(AbstractCheck.Meta):
-        abstract = False
-        swappable = swappable_setting('check', 'Check')
