+++ openwisp_monitoring/device/admin.py
@@ -28,7 +28,7 @@ Chart = load_model('monitoring', 'Chart')
 Device = load_model('config', 'Device')
 Metric = load_model('monitoring', 'Metric')
 Notification = load_model('openwisp_notifications', 'Notification')
-Check = load_model('check', 'Check')
+Check = load_model('monitoring', 'Check')
 
 
 class CheckInlineFormSet(BaseGenericInlineFormSet):
