+++ openwisp_monitoring/device/models.py
@@ -8,7 +8,7 @@ BaseDevice = load_model_patched('config', 'Device', require_ready=False)
 
 
 class DeviceData(AbstractDeviceData, BaseDevice):
-    checks = GenericRelation(get_model_name('check', 'Check'))
+    checks = GenericRelation(get_model_name('monitoring', 'Check'))
     metrics = GenericRelation(get_model_name('monitoring', 'Metric'))
 
     class Meta:
