+++ openwisp_monitoring/device/tasks.py
@@ -4,7 +4,7 @@ from celery import shared_task
 from django.core.exceptions import ObjectDoesNotExist
 from swapper import load_model
 
-from ..check.tasks import perform_check
+from ..monitoring.check.tasks import perform_check
 
 logger = logging.getLogger(__name__)
