+++ openwisp_monitoring/device/tests/test_admin.py
@@ -12,7 +12,7 @@ Chart = load_model('monitoring', 'Chart')
 Metric = load_model('monitoring', 'Metric')
 DeviceData = load_model('device_monitoring', 'DeviceData')
 User = get_user_model()
-Check = load_model('check', 'Check')
+Check = load_model('monitoring', 'Check')
 
 
 class TestAdmin(DeviceMonitoringTestCase):
