+++ openwisp_monitoring/monitoring/check/apps.py
@@ -3,7 +3,7 @@ from django.db.models.signals import post_save
 from django.utils.translation import ugettext_lazy as _
 from swapper import load_model
 
-from openwisp_monitoring.check import settings as app_settings
+from openwisp_monitoring.monitoring.check import settings as app_settings
 
 
 class CheckConfig(AppConfig):
