+++ openwisp_monitoring/monitoring/check/base/models.py
@@ -8,11 +8,11 @@ from django.utils.module_loading import import_string
 from django.utils.translation import gettext_lazy as _
 from jsonfield import JSONField
 
-from openwisp_monitoring.check import settings as app_settings
-from openwisp_monitoring.check.tasks import auto_create_config_check, auto_create_ping
+from openwisp_monitoring.monitoring.check import settings as app_settings
+from openwisp_monitoring.monitoring.check.tasks import auto_create_config_check, auto_create_ping
 from openwisp_utils.base import TimeStampedEditableModel
 
-from ...utils import transaction_on_commit
+from openwisp_monitoring.utils import transaction_on_commit
 
 
 class AbstractCheck(TimeStampedEditableModel):
