+++ openwisp_monitoring/monitoring/check/classes/config_applied.py
@@ -1,6 +1,6 @@
 from swapper import load_model
 
-from ...device.utils import SHORT_RP
+from openwisp_monitoring.device.utils import SHORT_RP
 from .base import BaseCheck
 
 AlertSettings = load_model('monitoring', 'AlertSettings')
