+++ openwisp_monitoring/monitoring/check/settings.py
