+++ openwisp_monitoring/monitoring/check/tasks.py
@@ -11,7 +11,7 @@ logger = logging.getLogger(__name__)
 
 
 def get_check_model():
-    return load_model('check', 'Check')
+    return load_model('monitoring', 'Check')
 
 
 @shared_task
