+++ tests/openwisp2/sample_check/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.utils.timezone import now
 
-from openwisp_monitoring.check.base.models import AbstractCheck
+from openwisp_monitoring.monitoring.check.base.models import AbstractCheck
 
 
 class Check(AbstractCheck):
