+++ tests/openwisp2/sample_check/tests.py
@@ -4,14 +4,14 @@ from django.test import TransactionTestCase
 from django.utils.timezone import now
 from swapper import load_model
 
-from openwisp_monitoring.check.classes import Ping
-from openwisp_monitoring.check.tests import _FPING_REACHABLE
-from openwisp_monitoring.check.tests.test_models import TestModels as BaseTestModels
-from openwisp_monitoring.check.tests.test_ping import TestPing as BaseTestPing
-from openwisp_monitoring.check.tests.test_utils import TestUtils as BaseTestUtils
+from openwisp_monitoring.monitoring.check.classes import Ping
+from openwisp_monitoring.monitoring.check.tests import _FPING_REACHABLE
+from openwisp_monitoring.monitoring.check.tests.test_models import TestModels as BaseTestModels
+from openwisp_monitoring.monitoring.check.tests.test_ping import TestPing as BaseTestPing
+from openwisp_monitoring.monitoring.check.tests.test_utils import TestUtils as BaseTestUtils
 from openwisp_monitoring.device.tests import TestDeviceMonitoringMixin
 
-Check = load_model('check', 'Check')
+Check = load_model('monitoring', 'Check')
 
 
 class TestUtils(BaseTestUtils, TestDeviceMonitoringMixin, TransactionTestCase):
